backend/robotservice/requirements.txt
Flask
flask_restful
python-dotenv
itsdangerous
psycopg2-binary
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
Flask-Migrate
gunicorn
Werkzeug
Flask-JWT-Extended
PyJWT
flask-cors
Flask-MQTT
pika